+++ flutter/lib/src/integrations/native_app_start_integration.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: invalid_use_of_internal_member
+
 import 'dart:async';
 
 import 'package:meta/meta.dart';
@@ -10,10 +12,13 @@ import '../event_processor/native_app_start_event_processor.dart';
 /// Integration which handles communication with native frameworks in order to
 /// enrich [SentryTransaction] objects with app start data for mobile vitals.
 class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
-  NativeAppStartIntegration(this._native, this._frameCallbackHandler);
+  NativeAppStartIntegration(this._native, this._frameCallbackHandler,
+      {Hub? hub})
+      : _hub = hub ?? HubAdapter();
 
   final SentryNative _native;
   final FrameCallbackHandler _frameCallbackHandler;
+  final Hub _hub;
 
   /// Timeout duration to wait for the app start info to be fetched.
   static const _timeoutDuration = Duration(seconds: 30);
@@ -55,13 +60,15 @@ class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
   @override
   void call(Hub hub, SentryFlutterOptions options) {
     if (isIntegrationTest) {
-      final appStartInfo = AppStartInfo(AppStartType.cold,
-          start: DateTime.now(),
-          end: DateTime.now().add(const Duration(milliseconds: 100)),
-          pluginRegistration:
-              DateTime.now().add(const Duration(milliseconds: 50)),
-          mainIsolateStart:
-              DateTime.now().add(const Duration(milliseconds: 60)));
+      final appStartInfo = AppStartInfo(
+        AppStartType.cold,
+        start: DateTime.now(),
+        end: DateTime.now().add(const Duration(milliseconds: 100)),
+        pluginRegistration:
+            DateTime.now().add(const Duration(milliseconds: 50)),
+        sentrySetupStart: DateTime.now().add(const Duration(milliseconds: 60)),
+        nativeSpanTimes: [],
+      );
       setAppStartInfo(appStartInfo);
       return;
     }
@@ -77,7 +84,12 @@ class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
         return;
       }
 
-      final mainIsolateStartDateTime = SentryFlutter.mainIsolateStartTime;
+      final sentrySetupStartDateTime = SentryFlutter.sentrySetupStartTime;
+      if (sentrySetupStartDateTime == null) {
+        setAppStartInfo(null);
+        return;
+      }
+
       final appStartDateTime = DateTime.fromMillisecondsSinceEpoch(
           nativeAppStart.appStartTime.toInt());
       final pluginRegistrationDateTime = DateTime.fromMillisecondsSinceEpoch(
@@ -86,7 +98,6 @@ class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
 
       if (options.autoAppStart) {
         // We only assign the current time if it's not already set - this is useful in tests
-        // ignore: invalid_use_of_internal_member
         _native.appStartEnd ??= options.clock();
         appStartEndDateTime = _native.appStartEnd;
 
@@ -106,14 +117,61 @@ class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
         }
       }
 
+      List<TimeSpan> nativeSpanTimes = [];
+      for (final entry in nativeAppStart.nativeSpanTimes.entries) {
+        try {
+          final startTimestampMs =
+              entry.value['startTimestampMsSinceEpoch'] as int;
+          final endTimestampMs =
+              entry.value['stopTimestampMsSinceEpoch'] as int;
+          nativeSpanTimes.add(TimeSpan(
+            start: DateTime.fromMillisecondsSinceEpoch(startTimestampMs),
+            end: DateTime.fromMillisecondsSinceEpoch(endTimestampMs),
+            description: entry.key as String,
+          ));
+        } catch (e) {
+          _hub.options.logger(
+              SentryLevel.warning, 'Failed to parse native span times: $e');
+          continue;
+        }
+      }
+
+      // We want to sort because the native spans are not guaranteed to be in order.
+      // Performance wise this won't affect us since the native span amount is very low.
+      nativeSpanTimes.sort((a, b) => a.start.compareTo(b.start));
+
       final appStartInfo = AppStartInfo(
           nativeAppStart.isColdStart ? AppStartType.cold : AppStartType.warm,
           start: appStartDateTime,
           end: appStartEndDateTime,
           pluginRegistration: pluginRegistrationDateTime,
-          mainIsolateStart: mainIsolateStartDateTime);
+          sentrySetupStart: sentrySetupStartDateTime,
+          nativeSpanTimes: nativeSpanTimes);
 
       setAppStartInfo(appStartInfo);
+
+      // When we don't have a SentryNavigatorObserver, a TTID transaction
+      // is not created therefore we need to create a transaction ourselves.
+      // We detect this by checking if the currentRouteName is null.
+      // This is a workaround since there is no api that tells us if
+      // the navigator observer exists and has been attached.
+      // The navigator observer also triggers much earlier so if it was attached
+      // it would have already set the routeName and the isCreated flag.
+      // The currentRouteName is always set during a didPush triggered
+      // by the navigator observer.
+      if (!SentryNavigatorObserver.isCreated &&
+          SentryNavigatorObserver.currentRouteName == null) {
+        const screenName = SentryNavigatorObserver.rootScreenName;
+        final transaction = hub.startTransaction(
+            screenName, SentrySpanOperations.uiLoad,
+            startTimestamp: appStartInfo.start);
+        final ttidSpan = transaction.startChild(
+            SentrySpanOperations.uiTimeToInitialDisplay,
+            description: '$screenName initial display',
+            startTimestamp: appStartInfo.start);
+        await ttidSpan.finish(endTimestamp: appStartInfo.end);
+        await transaction.finish(endTimestamp: appStartInfo.end);
+      }
     });
 
     options.addEventProcessor(NativeAppStartEventProcessor(_native, hub: hub));
@@ -129,19 +187,21 @@ class AppStartInfo {
     this.type, {
     required this.start,
     required this.pluginRegistration,
-    required this.mainIsolateStart,
+    required this.sentrySetupStart,
+    required this.nativeSpanTimes,
     this.end,
   });
 
   final AppStartType type;
   final DateTime start;
+  final List<TimeSpan> nativeSpanTimes;
 
   // We allow the end to be null, since it might be set at a later time
   // with setAppStartEnd when autoAppStart is disabled
   DateTime? end;
 
   final DateTime pluginRegistration;
-  final DateTime mainIsolateStart;
+  final DateTime sentrySetupStart;
 
   Duration? get duration => end?.difference(start);
 
@@ -160,6 +220,14 @@ class AppStartInfo {
   String get appStartTypeDescription =>
       type == AppStartType.cold ? 'Cold start' : 'Warm start';
   final pluginRegistrationDescription = 'App start to plugin registration';
-  final mainIsolateSetupDescription = 'Main isolate setup';
+  final sentrySetupDescription = 'Before Sentry Init Setup';
   final firstFrameRenderDescription = 'First frame render';
+}
+
+class TimeSpan {
+  TimeSpan({required this.start, required this.end, required this.description});
+
+  final DateTime start;
+  final DateTime end;
+  final String description;
 }